Moved the client code to lib folder.
Added an api to return size of checksum.
