+++ lib/Makefile.sub
@@ -44,6 +44,7 @@ nobase_motr_include_HEADERS += lib/arith.h \
                                lib/varr.h \
                                lib/varr_private.h \
                                lib/vec.h \
+                               lib/cksum.h \
                                lib/user_space/getopts.h \
                                lib/user_space/misc.h \
                                lib/user_space/mutex.h \
@@ -95,6 +96,7 @@ motr_libmotr_la_SOURCES += lib/assert.c \
                            lib/uuid.c \
                            lib/varr.c \
                            lib/vec.c \
+                           lib/cksum.c \
                            lib/user_space/finject_init.c \
                            lib/user_space/fs.c \
                            lib/user_space/memory.c \
